feat(ui): Implemented Notification header and metrics card
