+++ src/app/dashboard/system-management/notification-center/page.tsx
@@ -3,7 +3,5 @@
 import NotificationCenter from "@/pages/SystemManagement/NotificationCenter";
 
 export default function NotificationCenterPage() {
-  return (
-    <NotificationCenter/>
-  );
+  return <NotificationCenter />;
 }
